% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{absorp_ratio}
\alias{absorp_ratio}
\title{Absorption Ratio}
\usage{
absorp_ratio(xcor, n_pc = NULL)
}
\arguments{
\item{xcor}{correlation matrix}

\item{n_pc}{number of PCs for calculation, see details}
}
\description{
Absorption Ratio
}
\details{
From Kritzman, et al (2010), sum(1:n_pc) / sum(all pcs)
}
