% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{arc_to_dataframe}
\alias{arc_to_dataframe}
\title{Convert arctic data with columns as rownames into data.frame with
  Date column}
\usage{
arc_to_dataframe(dat)
}
\description{
Convert arctic data with columns as rownames into data.frame with
  Date column
}
\section{Fields}{

\describe{
\item{\code{dat}}{dataframe with dates as rownames}
}}

