% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_batting_avg}
\alias{calc_batting_avg}
\title{Calculate Batting Average}
\usage{
calc_batting_avg(x, mar = 0)
}
\arguments{
\item{x}{xts}

\item{mar}{sets place to sub-set up / down, default is `0`}
}
\description{
Calculate Batting Average
}
\details{
Average up day / Average down day, or week, month, etc
}
