% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_down_capture}
\alias{calc_down_capture}
\title{Calculate Downside Capture}
\usage{
calc_down_capture(mgr, bench, period, mar = 0)
}
\arguments{
\item{mgr}{xts object of manager / asset / security / etc to calculate downside capture}

\item{bench}{xts object of benchmark}

\item{mar}{sets place to sub-set up / down, default is `0`}
}
\value{
vector of downside capture
}
\description{
Calculate Downside Capture
}
\details{
Geometric return of manager / geometric return of benchmark when benchmark is up
}
