% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_down_vol}
\alias{calc_down_vol}
\title{Calculate Downside Volatiltiy}
\usage{
calc_down_vol(x, period, mar = 0)
}
\arguments{
\item{x}{xts object}

\item{period}{days, weeks, months, quarters to annualize volatiltiy}

\item{mar}{default `0`, sets min acceptable return to subset downside volatility}
}
\value{
vector of downside volatility
}
\description{
Calculate Downside Volatiltiy
}
\details{
Downside volatiltiy is the standard devation of returns for the sub-set of
the time-series below the `mar`
}
