% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_geo_ret}
\alias{calc_geo_ret}
\title{Calculate geometric return}
\usage{
calc_geo_ret(x, period)
}
\arguments{
\item{x}{xts object}

\item{period}{days, weeks, months, quarters to annualize return}
}
\value{
vector of geometric returns
}
\description{
Calculate geometric return
}
