% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_max_drawdown}
\alias{calc_max_drawdown}
\title{Calculate Worst Drawdown}
\usage{
calc_max_drawdown(x)
}
\arguments{
\item{x}{xts}
}
\value{
vector of worst drawdowns for each column in `x`
}
\description{
Calculate Worst Drawdown
}
