% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_uni_beta}
\alias{calc_uni_beta}
\title{Calculate Univariate Beta}
\usage{
calc_uni_beta(mgr, bench, rf)
}
\arguments{
\item{mgr}{xts of manager(s) returns}

\item{bench}{xts of benchmark returns}

\item{rf}{xts of risk-free returns}
}
\value{
vector of betas to benchmark
}
\description{
Calculate Univariate Beta
}
