% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_up_capture}
\alias{calc_up_capture}
\title{Calculate Upside Capture}
\usage{
calc_up_capture(mgr, bench, period, mar = 0)
}
\arguments{
\item{mgr}{xts object of manager / asset / security / etc to calculate upside capture}

\item{bench}{xts object of benchmark}

\item{mar}{sets place to sub-set up / down, default is `0`}
}
\value{
vector of upside capture
}
\description{
Calculate Upside Capture
}
\details{
Geometric return of manager / geometric return of benchmark when benchmark is up
}
