% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_vol}
\alias{calc_vol}
\title{Calculate Volatiltiy}
\usage{
calc_vol(x, period)
}
\arguments{
\item{x}{xts object}

\item{period}{days, weeks, months, quarters to annualize volatiltiy}
}
\value{
vector of volatility
}
\description{
Calculate Volatiltiy
}
