% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{calc_wgt_multiple}
\alias{calc_wgt_multiple}
\title{Calculate Weighted Harmonic Mean of a Ratio in a data.frame}
\usage{
calc_wgt_multiple(tbl_hold, x = "PE", w = "CapWgt")
}
\arguments{
\item{tbl_hold}{data.frame}

\item{x}{column name that contains the ratio}

\item{w}{column name that contains the weights}
}
\description{
Calculate Weighted Harmonic Mean of a Ratio in a data.frame
}
\details{
excludes missing and negative ratio and reconstitutes the weights
}
