% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{change_freq}
\alias{change_freq}
\title{Change time-series frequency}
\usage{
change_freq(x, period = "months", dtype = c("return", "price"))
}
\arguments{
\item{x}{xts object}

\item{period}{character string of the desired time-series periods}

\item{dtype}{character string of "return" or "price" to represent the data type}
}
\value{
xts object with new frequency
}
\description{
Change time-series frequency
}
