% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{check_tbl_hold}
\alias{check_tbl_hold}
\title{Check Holdings Table Specification}
\usage{
check_tbl_hold(tbl_hold)
}
\arguments{
\item{tbl_hold}{data.frame with holdings, need id, CapWgt, and TimeStamp}
}
\description{
Check Holdings Table Specification
}
