% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{clean_asset_bench_rf}
\alias{clean_asset_bench_rf}
\title{Handle combing and cleaning asset, rf, and benchmarks}
\usage{
clean_asset_bench_rf(x, b, rf = NULL)
}
\arguments{
\item{x}{xts object of asset(s)}

\item{b}{xts object of benchmark(s)}

\item{rf}{optional xts object of risk-free}
}
\description{
Handle combing and cleaning asset, rf, and benchmarks
}
\details{
Utility function to handle common task of needing to line up asset, bench,
  and rf returns. Function will truncate all time-series based on common
  start and end dates, and will clean returns and remove missing columns if
  NAs exceed 5% threshold.
}
