% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{cov_ewma}
\alias{cov_ewma}
\title{Exponentially Weighted Moving Average Covariance}
\usage{
cov_ewma(ret, lamda = NULL)
}
\arguments{
\item{ret}{xts of returns}

\item{lamda}{numeric value to scale weighted average calc}
}
\description{
Exponentially Weighted Moving Average Covariance
}
\note{
If lamda is left to default NULL it is calculated as 1 - 2 / a, where a
= the periodicity scaler (e.g., months = 12)
}
