% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imb_writer_fnc.R
\name{create_char_tbl}
\alias{create_char_tbl}
\title{Create Characteristics Table}
\usage{
create_char_tbl(rpt)
}
\arguments{
\item{rpt}{Reporter Object}

\item{freq}{Frequency of returns}
}
\description{
Create Characteristics Table
}
