% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imb_writer_fnc.R
\name{create_perf_tbl}
\alias{create_perf_tbl}
\title{Performance Table}
\usage{
create_perf_tbl(rpt, tm10, freq = "months")
}
\arguments{
\item{rpt}{Reporter Object}

\item{tm10}{t minus 10 years date to truncate return start}

\item{freq}{Frequency of returns}
}
\value{
flextable with formatted performance stats: alpha, beta, sharpe, TE,
  and up / down capture
}
\description{
Performance Table
}
