% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{cut_time}
\alias{cut_time}
\title{Truncate time-series with start and end dates}
\usage{
cut_time(x, date_start = NULL, date_end = NULL)
}
\arguments{
\item{x}{xts object}

\item{date_start}{optional beginning date to truncate time-series}

\item{date_end}{optional ending date to truncate time-series}
}
\description{
Truncate time-series with start and end dates
}
\details{
If neither date_start or date_end are entered the original
time series, x, will be returned. The date_start and date_end
parameters are inclusive, i.e., the date_start will be the first
date in the time-series.
}
\examples{
\dontrun{
data(etf_ret)
# cut time-series to start in Feb 1, 2010 and end on June 30, 2010
cut_time(etf_ret, '2010-02-01', '2010-06-30')
# leave beginning of time-serie as is, end on June 30, 2010
cut_time(etf_ret, NULL, '2010-06-30')
}
}
