% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{daily_spline}
\alias{daily_spline}
\title{Splice a monthly time-series with a daily time-series.}
\usage{
daily_spline(d_ret, m_ret, thresh = 1e-07)
}
\arguments{
\item{d_ret}{daily returns xts}

\item{m_ret}{monthly returns xts}

\item{thresh}{epsilon of difference between return series, see details}

\item{details}{The daily time-series is adjusted by a constant each month so it
its compounded return exactly equals (less epsilon) the compounded return
of the monthly time-series. A good use case is splicing monthly CTF
returns with daily estimates.}
}
\description{
Splice a monthly time-series with a daily time-series.
}
