% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{dataframe_to_arc}
\alias{dataframe_to_arc}
\title{Convert dataframe with Date column to arctic storage dataframe with
  dates in rownames}
\usage{
dataframe_to_arc(xdf)
}
\arguments{
\item{xdf}{data.frame with Date column (as first column xdf[, 1])}
}
\description{
Convert dataframe with Date column to arctic storage dataframe with
  dates in rownames
}
