% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{down_beta}
\alias{down_beta}
\title{Calculate Beta to benchmark when benchmark is down}
\usage{
down_beta(mgr, bench, rf, mar = 0)
}
\arguments{
\item{mgr}{xts of manager(s) returns}

\item{bench}{xts of benchmark returns}

\item{rf}{xts of risk-free returns}

\item{mar}{numeric minimal acceptable return for up / down split}
}
\description{
Calculate Beta to benchmark when benchmark is down
}
