% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_bd}
\alias{download_bd}
\title{Download Holdings from a Black Diamond Account}
\usage{
download_bd(account_id, api_keys, as_of = NULL)
}
\arguments{
\item{account_id}{external account ID from Black Diamond}

\item{api_keys}{list of API Keys}

\item{as_of}{as of date for holdings, leave \code{NULL} to default to
last NYSE trading day}
}
\value{
\code{data.frame} with holdings name, ticker, cusip, and value
}
\description{
Download Holdings from a Black Diamond Account
}
