% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_bd_batch_id}
\alias{download_bd_batch_id}
\title{Get Batch ID for bulk download}
\usage{
download_bd_batch_id(api_keys, as_of = NULL)
}
\arguments{
\item{api_keys}{list of API Keys}

\item{as_of}{optional date, defaults to last trading day, transactions
will default to last 7 days}
}
\value{
batch id to pass to \code{\link{download_bd_batch}} for bulk download
}
\description{
Get Batch ID for bulk download
}
