% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_fs_formula}
\alias{download_fs_formula}
\title{Download Factset formula API}
\usage{
download_fs_formula(
  api_keys,
  ids,
  formulas,
  type = c("ts", "cs"),
  flatn = TRUE
)
}
\arguments{
\item{api_keys}{list containing api keys}

\item{ids}{vector of ids to download}

\item{formulas}{formula to download}

\item{type}{"ts" for times-series or "cs" for cross-sectional}

\item{flatn}{boolean to add "flatten=Y" to api url}
}
\value{
json data
}
\description{
Download Factset formula API
}
