% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_fs_ra_ret}
\alias{download_fs_ra_ret}
\title{Factset Formula API: RA_RET}
\usage{
download_fs_ra_ret(id, api_keys, t_minus = 12, freq = "D")
}
\arguments{
\item{id}{character representing id (only supports one id)}

\item{t_minus}{number of months of desired time-series, numeric value}

\item{freq}{"D", or "M" for daily or monthly time-series}
}
\value{
xts time-series
}
\description{
Factset Formula API: RA_RET
}
\note{
For downloading returns of investments not on an exchange, e.g., mutual
  funds, indexes, DTC models, etc
}
\examples{
load("~/api_keys.RData")
download_fs_ra_ret(
  id = "GQETX",
  api_keys = api_keys,
  t_minus = 1,
  freq = "D"
)
}
