% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_sec}
\alias{download_sec}
\title{Download latest N-Port data from SEC EDGAR}
\usage{
download_sec(long_cik, short_cik, user_email)
}
\arguments{
\item{long_cik}{fund specific CIK}

\item{short_cik}{parent company CIK}

\item{user_email}{email address needs to be provided for scraping SEC data 
(no login or registration needed)}
}
\value{
\code{data.frame} with holdings and meta-data
}
\description{
Download latest N-Port data from SEC EDGAR
}
\examples{
# GMO Quality (GQEFX)
download_sec("S000004084", "772129", "asotolongo@diversifiedtrust.com")   
}
