% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{eom_cal_perf_dt}
\alias{eom_cal_perf_dt}
\title{Get Common Trailing Performance Dates}
\usage{
eom_cal_perf_dt(as_of = NULL, eom = TRUE)
}
\arguments{
\item{as_of}{starting date, default is end of last month}

\item{eom}{option to use end of month dates if working with monthly data,
default is TRUE}
}
\value{
vector of dates for common trailing periods, see details
}
\description{
Get Common Trailing Performance Dates
}
\details{
Will return DTD, MTD, QTD, 1 Yr, 3 Yr, 5 Yr, and 10 Yr trailing periods.
  If \code{eom} parameter is set to TRUE then DTD will be omitted.
}
