% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{excess_ret}
\alias{excess_ret}
\title{Calculate excess return (x - b)}
\usage{
excess_ret(x, b)
}
\arguments{
\item{x}{xts object to transform into excess return}

\item{b}{univariate xts object of benchmark or risk-free rate to subtract
from `x`}
}
\description{
Calculate excess return (x - b)
}
