% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{extract_list_null}
\alias{extract_list_null}
\title{Extract field from list checking for NULL}
\usage{
extract_list_null(x, nm)
}
\arguments{
\item{x}{list}

\item{nm}{field (string)}
}
\value{
NA if field is NULL or field value if not NULL
}
\description{
Extract field from list checking for NULL
}
