% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{factor_risk_wgt}
\alias{factor_risk_wgt}
\title{Calculate Factor Risk Weight}
\usage{
factor_risk_wgt(x_beta, x_cov, y_variance)
}
\arguments{
\item{x_beta}{factor betas from regression}

\item{x_cov}{covariance matrix of the regression factors}

\item{y_variance}{variance of the y variable}
}
\value{
a vector of the factor risk weights
}
\description{
Calculate Factor Risk Weight
}
\details{
Risk weights are the contribution to the total variance of the y variable,
i.e., the fund or asset to explain
}
