% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{fill_na_price}
\alias{fill_na_price}
\title{Fill Missing Price Data with Last Found Observation}
\usage{
fill_na_price(x)
}
\arguments{
\item{x}{xts}
}
\value{
xts with missing data filled
}
\description{
Fill Missing Price Data with Last Found Observation
}
\details{
For any missing data in a price time-series, found with `is.na`, the
last known price will be forward filled. E.g., for a monthly time-series
if Feb is missing than the Jan price will be filled in Feb.
}
