% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{find_drawdowns}
\alias{find_drawdowns}
\title{Find all the drawdowns of a time-series}
\usage{
find_drawdowns(x)
}
\arguments{
\item{x}{univariate xts, if more than one column is passed only the first
column will be used}
}
\value{
data.table with all the drawdowns, fields include StartDate,
   EndDate (recovery), TroughDate, Trough, DaysToTrough, DaysToRecover, and
   TotalDays
}
\description{
Find all the drawdowns of a time-series
}
