% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{flatten_fs_formula}
\alias{flatten_fs_formula}
\title{Utility to flatten json download}
\usage{
flatten_fs_formula(json)
}
\arguments{
\item{json}{download from \code{\link{download_fs_formula}}}
}
\value{
data.frame
}
\description{
Utility to flatten json download
}
