% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{flatten_fs_global_prices}
\alias{flatten_fs_global_prices}
\title{Utility to flatten json download}
\usage{
flatten_fs_global_prices(json, price = FALSE)
}
\arguments{
\item{json}{download from \code{\link{download_fs_global_prices}}}

\item{price}{boolean, "TRUE" if price data, "FALSE" if return data}
}
\value{
data.frame
}
\description{
Utility to flatten json download
}
