% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{freq_to_scaler}
\alias{freq_to_scaler}
\title{Change Character Frequency into Numeric Scaler}
\usage{
freq_to_scaler(period)
}
\arguments{
\item{period}{days, weeks, months, quarters, years}
}
\value{
corresponding numeric value, e.g., months = 12
}
\description{
Change Character Frequency into Numeric Scaler
}
