% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{latest_holdings}
\alias{latest_holdings}
\title{Truncate holdings table to most recent date}
\usage{
latest_holdings(tbl_hold)
}
\arguments{
\item{tbl_hold}{data.frame of holdings with date in TimeStamp field}
}
\value{
data.frame with holdings for the most recent date
}
\description{
Truncate holdings table to most recent date
}
