% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{mat_to_xts}
\alias{mat_to_xts}
\title{Convert matrix to xts object}
\usage{
mat_to_xts(x)
}
\arguments{
\item{x}{matrix with dates in first column}
}
\value{
xts
}
\description{
Convert matrix to xts object
}
