% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{merge_msl}
\alias{merge_msl}
\title{Merge MSL with holdings table}
\usage{
merge_msl(tbl_hold, tbl_msl, rm_dup_dates = TRUE)
}
\arguments{
\item{tbl_hold}{data.frame with portfolio holdings}

\item{tbl_msl}{master security list (data.frame)}
}
\value{
list with union, intersection, and missing tables
}
\description{
Merge MSL with holdings table
}
