% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{monthly_spline}
\alias{monthly_spline}
\title{Spline Quarterly Returns with Monthly Returns}
\usage{
monthly_spline(m_ret, q_ret, thresh = 1e-05)
}
\arguments{
\item{m_ret}{monthly returns xts}

\item{q_ret}{quarterly returns xts}

\item{thresh}{epsilon of difference between return series, see details}

\item{details}{The monthly time-series is adjusted by a constant each quarter so it
its compounded return exactly equals (less epsilon) the compounded return
of the quarterly time-series. A good use case is splicing quarterly
Private Equity TWR with the Russell 2000 monthly TWR.}
}
\description{
Spline Quarterly Returns with Monthly Returns
}
