% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{na_price}
\alias{na_price}
\title{Fill Missing Price Data for a Vector}
\usage{
na_price(x)
}
\arguments{
\item{x}{vector of price data}
}
\description{
Fill Missing Price Data for a Vector
}
\details{
see `fill_na_price`, which applies this function to an xts
}
