% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{port_beta}
\alias{port_beta}
\title{Beta between two portfolios}
\usage{
port_beta(wgt_vec_1, wgt_vec_2, cov_mat)
}
\arguments{
\item{wgt_vec_1}{weight vector of 1st portfolio}

\item{wgt_vec_2}{weight vector of 2nd portfolio}

\item{cov_mat}{union covariance of wgt_vec_1 and wgt_vec_2}
}
\description{
Beta between two portfolios
}
