% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{port_ret}
\alias{port_ret}
\title{Portfolio Return}
\usage{
port_ret(x, mu_vec)
}
\arguments{
\item{x}{vector of weights}

\item{mu_vec}{vector of expected or historic means}
}
\description{
Portfolio Return
}
