% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{Portfolio}
\alias{Portfolio}
\title{Portfolio Object}
\description{
Structure to contain and wrangle portfolio holdings and returns
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tbl_hold}}{holdings table}

\item{\code{tbl_miss}}{holdings that are missing from MSL}

\item{\code{tbl_msl}}{table containing master security list}

\item{\code{ac}}{ArcticDB datastore}

\item{\code{rebal}}{Rebalance object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Portfolio-new}{\code{Portfolio$new()}}
\item \href{#method-Portfolio-check_ac}{\code{Portfolio$check_ac()}}
\item \href{#method-Portfolio-check_tbl_hold}{\code{Portfolio$check_tbl_hold()}}
\item \href{#method-Portfolio-merge_msl}{\code{Portfolio$merge_msl()}}
\item \href{#method-Portfolio-drill_down}{\code{Portfolio$drill_down()}}
\item \href{#method-Portfolio-get_fina_data}{\code{Portfolio$get_fina_data()}}
\item \href{#method-Portfolio-get_sector_data}{\code{Portfolio$get_sector_data()}}
\item \href{#method-Portfolio-get_country_data}{\code{Portfolio$get_country_data()}}
\item \href{#method-Portfolio-read_ret}{\code{Portfolio$read_ret()}}
\item \href{#method-Portfolio-read_rebal_wgt}{\code{Portfolio$read_rebal_wgt()}}
\item \href{#method-Portfolio-init_rebal}{\code{Portfolio$init_rebal()}}
\item \href{#method-Portfolio-clone}{\code{Portfolio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-new"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-new}{}}}
\subsection{Method \code{new()}}{
Create new Portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$new(ac, tbl_hold, name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ac}}{ArcticDB datastore from Database Object}

\item{\code{tbl_hold}}{holdings table, see details}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
tbl_hold is gathered with the Database Object method $get_hold().
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-check_ac"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-check_ac}{}}}
\subsection{Method \code{check_ac()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$check_ac()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-check_tbl_hold"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-check_tbl_hold}{}}}
\subsection{Method \code{check_tbl_hold()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$check_tbl_hold()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-merge_msl"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-merge_msl}{}}}
\subsection{Method \code{merge_msl()}}{
Merge MSL with Holdings Table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$merge_msl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-drill_down"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-drill_down}{}}}
\subsection{Method \code{drill_down()}}{
Drill down to underlying holdings of funds / CTFs / models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$drill_down()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_fina_data"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_fina_data}{}}}
\subsection{Method \code{get_fina_data()}}{
Get Company
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_fina_data(xsymbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xsymbols}}{leave blank for all data, or specify specific data, e.g.,
`"PE"`, `"PB"`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_sector_data"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_sector_data}{}}}
\subsection{Method \code{get_sector_data()}}{
Get Sector Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_sector_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_country_data"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_country_data}{}}}
\subsection{Method \code{get_country_data()}}{
Get Country Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_country_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-read_ret"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-read_ret}{}}}
\subsection{Method \code{read_ret()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$read_ret()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-read_rebal_wgt"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-read_rebal_wgt}{}}}
\subsection{Method \code{read_rebal_wgt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$read_rebal_wgt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-init_rebal"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-init_rebal}{}}}
\subsection{Method \code{init_rebal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$init_rebal(rebal_freq = "M", asset_freq = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
