% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{price_to_ret}
\alias{price_to_ret}
\title{Convert price time-series to a return}
\usage{
price_to_ret(x)
}
\arguments{
\item{x}{xts of prices}
}
\value{
xts of returns
}
\description{
Convert price time-series to a return
}
\details{
Return is defined by period over period change
}
