% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{rbind_tx}
\alias{rbind_tx}
\title{Utility to row bind transaction data}
\usage{
rbind_tx(old, new)
}
\arguments{
\item{old}{existing data.frame of transactions}

\item{new}{newly downloaded transactions}
}
\value{
combo of new and old, removes duplicated values for new from old
}
\description{
Utility to row bind transaction data
}
\seealso{
used as part of \code{\link{handle_bd_batch}}
}
