% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{read_hold}
\alias{read_hold}
\title{Read underlying holdings}
\usage{
read_hold(ac, dtc_name, latest = TRUE)
}
\arguments{
\item{ac}{ArcticDB structure}

\item{dtc_name}{dtc id of security, string}

\item{latest}{boolean default is TRUE for only return most recent holdings,
FALSE will return a time-series of holdings over time}
}
\value{
data.frame with holdings values, weights, and meta-data
}
\description{
Read underlying holdings
}
