% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{read_macro_wb}
\alias{read_macro_wb}
\title{Read Macro Select Model Workbook}
\usage{
read_macro_wb(wb, idx_nm)
}
\arguments{
\item{wb}{full file name of workbook}

\item{fact_nm}{string vector representing the factor names}

\item{indx_nm}{name of index, e.g., "Russell 3000"}
}
\value{
data.frame
}
\description{
Read Macro Select Model Workbook
}
