% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_private_xts}
\alias{read_private_xts}
\title{Read Private I Excel Export}
\usage{
read_private_xts(file_nm, field_nm)
}
\arguments{
\item{file_nm}{full file name of workbook}

\item{field_nm}{column name for returns}
}
\value{
xts object of returns
}
\description{
Read Private I Excel Export
}
\examples{
read_private_xts(
  file_nm = "N:/Investment Team/DATABASES/CustomRet/PE-Downloads/PrivateEquity.xlsx",
  field_nm = "Private Equity"
)
}
