% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{read_ret}
\alias{read_ret}
\title{Read Returns}
\usage{
read_ret(ids, ac, tbl_msl = NULL)
}
\arguments{
\item{ids}{ticker, cusip, lei, dtc name, or identifier}

\item{ac}{arcticdb datastore}
}
\value{
xts of time-series of ids, will warn if some ids are not found
}
\description{
Read Returns
}
