% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{read_xts}
\alias{read_xts}
\title{Read excel time-series}
\usage{
read_xts(wb, sht = 1, skip = 0)
}
\arguments{
\item{wb}{workbook full file name, e.g.,
'C:/users/asotolongo/documents/wb.xslx'}

\item{sht}{worksheet number or name}

\item{skip}{number of header rows to skip, default is 0}
}
\description{
Read excel time-series
}
\details{
See `readxl::read_excel` for more info. Format of workbook needs to have
  date column in first row.
}
