% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rebal.R
\name{Rebal}
\alias{Rebal}
\title{Rebal Object}
\description{
Object to handle rebalancing asset returns into a portfolio time-series of
  returns. Rebalance weights can either be a vector or xts time-series of
  weights. Helper function can add periodic (e.g., monthly) rebalanc weights
  based on the previous specified weights.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{column name for returns after rebal}

\item{\code{rebal_wgt}}{rebalance weights}

\item{\code{asset_wgt}}{asset weights}

\item{\code{asset_ret}}{asset returns}

\item{\code{asset_freq}}{frequency of asset returns}

\item{\code{rebal_freq}}{frequency of rebalance}

\item{\code{rebal_ret}}{portfolio returns from}

\item{\code{ctr_mat}}{contribution to return matrix}

\item{\code{asset_ret_start}}{inception of asset returns}

\item{\code{asset_ret_end}}{last date of asset returns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Rebal-new}{\code{Rebal$new()}}
\item \href{#method-Rebal-wrangle_rebal_vect}{\code{Rebal$wrangle_rebal_vect()}}
\item \href{#method-Rebal-align_rebal_wgt}{\code{Rebal$align_rebal_wgt()}}
\item \href{#method-Rebal-fill_rebal_dates}{\code{Rebal$fill_rebal_dates()}}
\item \href{#method-Rebal-rebal}{\code{Rebal$rebal()}}
\item \href{#method-Rebal-clone}{\code{Rebal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rebal-new"></a>}}
\if{latex}{\out{\hypertarget{method-Rebal-new}{}}}
\subsection{Method \code{new()}}{
Create new rebal object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rebal$new(
  rebal_wgt,
  asset_ret,
  asset_freq = "D",
  rebal_freq = "M",
  name = "PortRebal"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rebal_wgt}}{vector or xts of rebalance weights}

\item{\code{asset_ret}}{xts of asset returns}

\item{\code{asset_freq}}{character for asset_ret frequency, see below}

\item{\code{rebal_freq}}{character: "D", "M", "Q", "A", "BH" for daily, monthly,
quarterly, annual, or buy and hold}

\item{\code{name}}{column name for rebalance returns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rebal-wrangle_rebal_vect"></a>}}
\if{latex}{\out{\hypertarget{method-Rebal-wrangle_rebal_vect}{}}}
\subsection{Method \code{wrangle_rebal_vect()}}{
change vector of weights into xts with start date of asset
  returns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rebal$wrangle_rebal_vect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rebal-align_rebal_wgt"></a>}}
\if{latex}{\out{\hypertarget{method-Rebal-align_rebal_wgt}{}}}
\subsection{Method \code{align_rebal_wgt()}}{
match columns of rebal weights and asset returns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rebal$align_rebal_wgt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rebal-fill_rebal_dates"></a>}}
\if{latex}{\out{\hypertarget{method-Rebal-fill_rebal_dates}{}}}
\subsection{Method \code{fill_rebal_dates()}}{
fill in periodic rebalance dates if frequency is specified
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rebal$fill_rebal_dates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rebal-rebal"></a>}}
\if{latex}{\out{\hypertarget{method-Rebal-rebal}{}}}
\subsection{Method \code{rebal()}}{
execute rebalance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rebal$rebal(sum_to_1 = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sum_to_1}}{boolean to force rebalance weights to equal 100% for each
period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rebal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Rebal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rebal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
