% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{refresh_bd_key}
\alias{refresh_bd_key}
\title{Refresh Black Diamond Token for API Authorization}
\usage{
refresh_bd_key(api_keys, save_to_n = FALSE, save_local = FALSE)
}
\arguments{
\item{api_keys}{list of API Keys}

\item{save_to_n}{if on DTC network option to save to N: drive}

\item{save_local}{save to local directory, e.g., "~/"}
}
\description{
Refresh Black Diamond Token for API Authorization
}
