% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_returns.R
\name{ret_ctf_monthly}
\alias{ret_ctf_monthly}
\title{Update CTF Returns from Factset}
\usage{
ret_ctf_monthly(ac, tbl_msl, api_keys, ids = NULL, t_minus_m = 1)
}
\arguments{
\item{ac}{ArcticDB datastore, see create_arctic()}

\item{tbl_msl}{master security list data.frame}

\item{api_keys}{list with keys}

\item{ids}{leave `NULL` to update all CTFs or enter a vector of ids to
only update specific CTFs}

\item{t_minus_m}{integer to indicate how many months back to download new
data from}
}
\description{
Update CTF Returns from Factset
}
