% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_returns.R
\name{ret_index}
\alias{ret_index}
\title{Update index returns from Factset}
\usage{
ret_index(ac, tbl_msl, api_keys, ids = NULL, t_minus_m = 1)
}
\arguments{
\item{ac}{ArcticDB datastore, see create_arctic()}

\item{tbl_msl}{master security list data.frame}

\item{api_keys}{list with keys}

\item{ids}{leave NULL to update all indexes, or enter specific index
ids to update, note ids will have to first exist in the Master Security
List `tbl_msl`}

\item{t_minus_m}{how many months to download}
}
\description{
Update index returns from Factset
}
