% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{ret_to_price}
\alias{ret_to_price}
\title{Convert returns to a price index}
\usage{
ret_to_price(x)
}
\arguments{
\item{x}{xts of returns}
}
\value{
a price index calculating from \code{x} and a initial value of 1
}
\description{
Convert returns to a price index
}
