% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{risk_cluster_wgt}
\alias{risk_cluster_wgt}
\title{Calculate risk cluster weights}
\usage{
risk_cluster_wgt(hc, vol, k = 2)
}
\arguments{
\item{hc}{tree from hclust output}

\item{vol}{volatility of assets used in hc calculation}

\item{k}{number of clusters}
}
\description{
Calculate risk cluster weights
}
