% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{rm_na_col}
\alias{rm_na_col}
\title{Remove columns in return time-series with too many missing values}
\usage{
rm_na_col(x, eps = 0.05, rpl = 0)
}
\arguments{
\item{x}{xts object}

\item{eps}{portion of data, enter 0.05 for 5%, at which to delete the column
if the missing value % of the column length exceeds. See details for more info.}

\item{rpl}{value to replace missing values, `0` by default}
}
\value{
xts with bad columns removed. Missing values are `NA` found by the
`is.na` function.
}
\description{
Remove columns in return time-series with too many missing values
}
\details{
If the the column length is 10 and the `eps` = 0.05, then if the missing values
in a column are greater than 2 (10 * 0.05 = 2) then remove the column. Any
missing data in a column that was less than the `eps` will be replaced with zero.
}
