% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{roll_ret}
\alias{roll_ret}
\title{Rolling returns}
\usage{
roll_ret(x, n, b = NULL, period = "days")
}
\arguments{
\item{x}{xts of returns}

\item{n}{number of periods for rolling calc window}

\item{b}{optional benchmark to transform to active return}

\item{period}{frequency, default is days}
}
\description{
Rolling returns
}
