% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{set_plot_col}
\alias{set_plot_col}
\title{Set n number of assets to plot to DTC color palette}
\usage{
set_plot_col(n)
}
\arguments{
\item{n}{number of assets / items to plot}
}
\value{
list of colors
}
\description{
Set n number of assets to plot to DTC color palette
}
\details{
if the number of assets exceed the number of colors in the DTC
  palette, 1:n colors will be generated based on numbers
}
