% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{space_ids}
\alias{space_ids}
\title{Space Out ids in multiple iterations}
\usage{
space_ids(ids, max_ids = 50)
}
\arguments{
\item{ids}{vector of ids}

\item{max_ids}{max number of ids per iteration}
}
\value{
iteration index
}
\description{
Space Out ids in multiple iterations
}
\details{
utility function for factset download that limits number of ids per
 api call
}
