% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{test_diff}
\alias{test_diff}
\title{Utility function for testing the difference of the product of A and B}
\usage{
test_diff(a, b)
}
\description{
Utility function for testing the difference of the product of A and B
}
