% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{unsmooth_ret}
\alias{unsmooth_ret}
\title{Unsmooth serially auto-correlated returns}
\usage{
unsmooth_ret(x)
}
\arguments{
\item{x}{return xts to unsmooth}
}
\description{
Unsmooth serially auto-correlated returns
}
