% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{unzip_bd_batch}
\alias{unzip_bd_batch}
\title{Unzip Response from BlackDiamond Batch Download}
\usage{
unzip_bd_batch(resp)
}
\arguments{
\item{resp}{response returned by \code{\link{download_bd_batch}}}
}
\value{
bulk data from the json download nested in a list
}
\description{
Unzip Response from BlackDiamond Batch Download
}
\seealso{
\code{\link{download_bd_batch}} for downloading the bulk data
}
