% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{upload_ctf_monthly}
\alias{upload_ctf_monthly}
\title{Upload Month End CTF Reconciled Returns into the Database}
\usage{
upload_ctf_monthly(ac, xl_path = NULL, skip = 4)
}
\arguments{
\item{ac}{arcticDB datastore}

\item{xl_path}{filepath of excel workbook with returns, leave `NULL` for
default}

\item{skip}{leading header rows to skip, default is `4`}
}
\value{
no data is returned, the entire history is read and stored in
  arcticDB
}
\description{
Upload Month End CTF Reconciled Returns into the Database
}
