% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{us_trading_days}
\alias{us_trading_days}
\title{Get US Trading Days based on NYSE Holidays}
\usage{
us_trading_days(date_start = NULL, date_end = NULL)
}
\arguments{
\item{date_start}{beginning date of sequence}

\item{date_end}{last date of sequence}
}
\value{
sequence of trading days
}
\description{
Get US Trading Days based on NYSE Holidays
}
