% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{viz_roll_beta}
\alias{viz_roll_beta}
\title{Plot Rolling Beta}
\usage{
viz_roll_beta(x, b, rf, n, freq)
}
\arguments{
\item{x}{xts object of assets to calculate beta for}

\item{b}{xts object of one benchmark to calculate beta to}

\item{rf}{xts object of risk-free rate}

\item{n}{number of rolling periods to use in the calculation}

\item{freq}{string to represent frequency for chart title}
}
\description{
Plot Rolling Beta
}
