% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{viz_roll_vol}
\alias{viz_roll_vol}
\title{Plot Rolling Volatility}
\usage{
viz_roll_vol(x, n, freq)
}
\arguments{
\item{x}{xts object with time-series of returns}

\item{n}{number of periods for rolling window}

\item{freq}{string to represent frequency for scaling}
}
\description{
Plot Rolling Volatility
}
