% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{viz_trade_off}
\alias{viz_trade_off}
\title{Plot Risk / Reward Trade-offs}
\usage{
viz_trade_off(
  x,
  b = NULL,
  risk_type = c("vol", "down-vol", "drawdown"),
  period = "days"
)
}
\arguments{
\item{x}{xts of asset returns}

\item{b}{optional xts of benchmark returns for relative trade-off}

\item{risk_type}{character to specify vol, down-vol, or drawdown, if b
is entered then TE, down-TE, and underperformance are calculated}

\item{period}{character representing frequency to annualized vol}
}
\description{
Plot Risk / Reward Trade-offs
}
