% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{viz_wealth_index}
\alias{viz_wealth_index}
\title{Plot Cumulative Returns}
\usage{
viz_wealth_index(x, init_val = 100)
}
\arguments{
\item{x}{xts object with time-series of returns}

\item{init_val}{initial starting value of wealth}
}
\description{
Plot Cumulative Returns
}
