% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{wgt_harmonic_mean}
\alias{wgt_harmonic_mean}
\title{Calculate Weighted Harmonic Mean}
\usage{
wgt_harmonic_mean(w, x)
}
\arguments{
\item{w}{weight vector}

\item{x}{data vector}
}
\description{
Calculate Weighted Harmonic Mean
}
