% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imb_bond_fund.R
\name{write_bond}
\alias{write_bond}
\title{Write IMB Bond Slide}
\usage{
write_bond(
  pres,
  rpt,
  dict,
  descr,
  locater,
  slide_title,
  tm10,
  pie_type = c("Quality", "Sector")
)
}
\arguments{
\item{pres}{presentation object from officer}

\item{rpt}{report object}

\item{dict}{dictionary table from excel}

\item{descr}{description table from excel}

\item{slide_title}{string for title}

\item{tm10}{t-minus 10 years date}

\item{pie_type}{plot quality or sector pie chart}

\item{locator}{list with chart and table locations}
}
\value{
pres with added slide
}
\description{
Write IMB Bond Slide
}
