% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imb_equity_ctf.R
\name{write_equity}
\alias{write_equity}
\title{Write IMB Equity Slide}
\usage{
write_equity(pres, rpt, dict, descr, locater, slide_title, tm10)
}
\arguments{
\item{pres}{presentation object from officer}

\item{rpt}{report object}

\item{dict}{dictionary table from excel}

\item{descr}{description table from excel}

\item{slide_title}{string for title}

\item{tm10}{t-minus 10 years date}

\item{locator}{list with chart and table locations}
}
\value{
pres with added slide
}
\description{
Write IMB Equity Slide
}
