% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{xts_cbind_inter}
\alias{xts_cbind_inter}
\title{Column bind the intersection of two xts objects}
\usage{
xts_cbind_inter(x, y)
}
\arguments{
\item{x}{xts object}

\item{y}{xts object}
}
\value{
list containing intersection of returns and columns that were removed
  due to too many missing values (if any)
}
\description{
Column bind the intersection of two xts objects
}
\details{
The first common start date and last common end date are used to find the
interection of returns. The returns are also cleaned for missing columns.
See rm_na_col for more info.
}
