% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{xts_rbind}
\alias{xts_rbind}
\title{Row bind xts}
\usage{
xts_rbind(new, old, is_xts = TRUE)
}
\arguments{
\item{new}{older time-series}

\item{old}{newer time-series}

\item{is_xts}{boolean, if set to FALSE, assumes new and old are data.frames
with `Date` column}
}
\description{
Row bind xts
}
\details{
For any overlapping dates, the old xts will be overwritten by the new xts.
The function will align the intersection of column names for the old and new
xts objects. For any column names that do not intersect NAs will be added
to the new or old xts depending on which xts is missing the column name(s).
}
