% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{xts_to_dataframe}
\alias{xts_to_dataframe}
\title{Convert xts object to a data.frame}
\usage{
xts_to_dataframe(x)
}
\arguments{
\item{x}{xts object}
}
\value{
data.frame with Dates in the first column
}
\description{
Convert xts object to a data.frame
}
