% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R, R/util_time_series.R
\name{clean_asset_bench_rf}
\alias{clean_asset_bench_rf}
\title{Handle combing and cleaning asset, rf, and benchmarks}
\usage{
clean_asset_bench_rf(
  x,
  b,
  rf = NULL,
  freq = NULL,
  date_start = NULL,
  date_end = NULL
)

clean_asset_bench_rf(
  x,
  b,
  rf = NULL,
  freq = NULL,
  date_start = NULL,
  date_end = NULL
)
}
\arguments{
\item{x}{asset time-series, xts, multiple columns accepted}

\item{b}{benchmark time-series, xts, mutliple columns accepted}

\item{rf}{risk-free time-series, optional, one column accepted}
}
\value{
list with `x`, `b`, and `rf` aligned for common dates
}
\description{
Handle combing and cleaning asset, rf, and benchmarks

Get Intersection of Asset, Benchmark, and Risk-free time-series
}
\details{
Utility function to handle common task of needing to line up asset, bench,
  and rf returns. Function will truncate all time-series based on common
  start and end dates, and will clean returns and remove missing columns if
  NAs exceed 5% threshold.
}
