% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imb_writer_fnc.R
\name{create_trail_perf_tbl}
\alias{create_trail_perf_tbl}
\title{Trailing Periodic Performance}
\usage{
create_trail_perf_tbl(rpt, freq = "months", p_or = NULL, b_or = NULL)
}
\arguments{
\item{rpt}{Reporter Object}

\item{freq}{Frequency of returns}
}
\description{
Trailing Periodic Performance
}
