% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imb_writer_fnc.R
\name{create_wealth_cht}
\alias{create_wealth_cht}
\title{Create Cumulative Return Chart}
\usage{
create_wealth_cht(rpt, tm10, freq = "M")
}
\arguments{
\item{rpt}{Reporter Object}

\item{tm10}{t minus 10 years date to truncate return start}

\item{freq}{return frequency}
}
\description{
Create Cumulative Return Chart
}
