% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{Database}
\alias{Database}
\title{Database Object}
\description{
Download and store holdings and price time-series data in s3 from various
  sources such as factset, blackdiamond, excel
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_keys}}{of api keys}

\item{\code{bucket}}{aws s3 bucket}

\item{\code{ac}}{ArcticDB object store}

\item{\code{tbl_msl}}{master security list: table of meta data for all investments}

\item{\code{tbl_cust}}{custodian list: table of meta data for CTFs and SMAs}

\item{\code{tbl_sec}}{SEC list: table of meta data for mutual funds and ETFs}

\item{\code{tbl_xl_mod}}{excel model list: table of meta data for model portfolios}

\item{\code{tbl_hold_field}}{holdings fields: table of fields / columns for
holdings data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Database-new}{\code{Database$new()}}
\item \href{#method-Database-check_api_keys}{\code{Database$check_api_keys()}}
\item \href{#method-Database-ret_index}{\code{Database$ret_index()}}
\item \href{#method-Database-ret_etf}{\code{Database$ret_etf()}}
\item \href{#method-Database-ret_mutual_fund}{\code{Database$ret_mutual_fund()}}
\item \href{#method-Database-ret_private_index}{\code{Database$ret_private_index()}}
\item \href{#method-Database-ret_ctf_monthly}{\code{Database$ret_ctf_monthly()}}
\item \href{#method-Database-read_ret_ticker}{\code{Database$read_ret_ticker()}}
\item \href{#method-Database-hold_sec}{\code{Database$hold_sec()}}
\item \href{#method-Database-hold_cust}{\code{Database$hold_cust()}}
\item \href{#method-Database-read_hold}{\code{Database$read_hold()}}
\item \href{#method-Database-download_fundamental_data}{\code{Database$download_fundamental_data()}}
\item \href{#method-Database-download_sectors}{\code{Database$download_sectors()}}
\item \href{#method-Database-update_ps_macro_select}{\code{Database$update_ps_macro_select()}}
\item \href{#method-Database-clone}{\code{Database$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-new"></a>}}
\if{latex}{\out{\hypertarget{method-Database-new}{}}}
\subsection{Method \code{new()}}{
Create a db object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$new(api_keys, py_loc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_keys}}{list of api keys or `.RData` file location to load the list
of keys}

\item{\code{py_loc}}{optional file path where python is installed to be used
with `reticulate`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-check_api_keys"></a>}}
\if{latex}{\out{\hypertarget{method-Database-check_api_keys}{}}}
\subsection{Method \code{check_api_keys()}}{
Check if api keys are properly specified
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$check_api_keys(api_keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_keys}}{list of api_keys}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-ret_index"></a>}}
\if{latex}{\out{\hypertarget{method-Database-ret_index}{}}}
\subsection{Method \code{ret_index()}}{
Update index returns from Factset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$ret_index(ids = NULL, t_minus = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{leave NULL to update all indexes, or enter specific index
ids to update, note ids will have to first exist in the Master Security
List `tbl_msl`}

\item{\code{t_minus}}{how many months to download}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-ret_etf"></a>}}
\if{latex}{\out{\hypertarget{method-Database-ret_etf}{}}}
\subsection{Method \code{ret_etf()}}{
Update ETF returns from Factset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$ret_etf(ids = NULL, date_start = NULL, date_end = Sys.Date() - 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{leave `NULL` to udpate all ETFs in the Master Security List,
or enter a vector to only update specific ETFs}

\item{\code{date_start}}{beginning date for update, if left `NULL` will
default to the last date of the existing data}

\item{\code{date_end}}{ending date for update, by default is yesterday}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-ret_mutual_fund"></a>}}
\if{latex}{\out{\hypertarget{method-Database-ret_mutual_fund}{}}}
\subsection{Method \code{ret_mutual_fund()}}{
Update index returns from Factset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$ret_mutual_fund(ids = NULL, days_back = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{leave NULL to update all mutual funds, or enter specific
mutual fund, note if ids are entered they must be in the MSL}

\item{\code{days_back}}{how many days to download}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-ret_private_index"></a>}}
\if{latex}{\out{\hypertarget{method-Database-ret_private_index}{}}}
\subsection{Method \code{ret_private_index()}}{
Update Private Asset Indexes from Excel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$ret_private_index()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-ret_ctf_monthly"></a>}}
\if{latex}{\out{\hypertarget{method-Database-ret_ctf_monthly}{}}}
\subsection{Method \code{ret_ctf_monthly()}}{
Update CTF Returns from Factset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$ret_ctf_monthly(ids = NULL, t_minus = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{leave `NULL` to update all CTFs or enter a vector of ids to
only update specific CTFs}

\item{\code{t_minus}}{integer to indicate how many months back to download new
data from}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-read_ret_ticker"></a>}}
\if{latex}{\out{\hypertarget{method-Database-read_ret_ticker}{}}}
\subsection{Method \code{read_ret_ticker()}}{
Read Returns by Ticker
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$read_ret_ticker(ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{tickers to read in}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-hold_sec"></a>}}
\if{latex}{\out{\hypertarget{method-Database-hold_sec}{}}}
\subsection{Method \code{hold_sec()}}{
Download holdings from SEC EDGAR Database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$hold_sec(
  dtc_name = NULL,
  user_email = "asotolongo@diversifiedtrust.com",
  save_to_db = TRUE,
  return_data = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dtc_name}}{leave `NULL` to download all, or enter a vector of
dtc_names to download specific funds}

\item{\code{user_email}}{need to provide an email address to download}

\item{\code{save_to_db}}{save data to DTC's database}

\item{\code{return_data}}{return data.frame of holdings}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-hold_cust"></a>}}
\if{latex}{\out{\hypertarget{method-Database-hold_cust}{}}}
\subsection{Method \code{hold_cust()}}{
Download account data from custodian, currently done via
  BlackDiamond
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$hold_cust(dtc_name = NULL, save_to_db = TRUE, return_data = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dtc_name}}{leave `NULL` to download all, or enter a vector of
dtc_names to download specific funds}

\item{\code{save_to_db}}{save data to DTC's database}

\item{\code{return_data}}{return data.frame of holdings}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-read_hold"></a>}}
\if{latex}{\out{\hypertarget{method-Database-read_hold}{}}}
\subsection{Method \code{read_hold()}}{
Read Holdings Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$read_hold(dtc_name, latest = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dtc_name}}{DtcName field in MSL to pull holdings}

\item{\code{latest}}{option to truncate to most recent holdings}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-download_fundamental_data"></a>}}
\if{latex}{\out{\hypertarget{method-Database-download_fundamental_data}{}}}
\subsection{Method \code{download_fundamental_data()}}{
Download equity financial data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$download_fundamental_data(
  ids = NULL,
  yrs_back = 1,
  dtype = c("PE", "PB", "PFCF", "DY", "ROE", "MCAP")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{leave `NULL` to get for all stocks, or enter specific ids}

\item{\code{yrs_back}}{how many years of data to pull}

\item{\code{dtype}}{data type: one of PE, PB, PFCF, DY, ROE, and MCAP}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-download_sectors"></a>}}
\if{latex}{\out{\hypertarget{method-Database-download_sectors}{}}}
\subsection{Method \code{download_sectors()}}{
Download Sector Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$download_sectors(ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{specify ids or leave `NULL` to gather for all stocks in MSL}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sector data are gathered from Factset, mapped to GICS, and from 
Piper Sandler Macro Downloads
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_ps_macro_select"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_ps_macro_select}{}}}
\subsection{Method \code{update_ps_macro_select()}}{
Download Macro Select Workbook and Save to Library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_ps_macro_select(wb, is_us = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wb}}{file location of workbook}

\item{\code{is_us}}{TRUE for Russell 3000, FALSE for MSCI ACWI}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Database-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
