% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_bd_batch}
\alias{download_bd_batch}
\title{BlackDiamond Batch Download}
\usage{
download_bd_batch(api_keys, batch_id)
}
\arguments{
\item{api_keys}{list of API Keys}

\item{batch_id}{from `download_bd_batch_id`}
}
\value{
response with raw data
}
\description{
BlackDiamond Batch Download
}
