% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_hfr_csv}
\alias{read_hfr_csv}
\title{Read HFR Returns CSV Export}
\usage{
read_hfr_csv(file_nm)
}
\arguments{
\item{file_nm}{file name of csv file, e.g., 
"C:/users/asotolongo/Downloads/HFR Indices.csv"}
}
\value{
xts object of returns
}
\description{
Read HFR Returns CSV Export
}
\examples{
dat <- read_hfr_csv("C:/users/asotolongo/Downloads/HFR Indices.csv")
tail(dat[, 1:5])
}
