% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporter.R
\name{Reporter}
\alias{Reporter}
\title{Reporter Object}
\description{
Structure to wrangle portfolios and an optional benchmark into reports
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{port}}{list of Portfolio Object(s)}

\item{\code{bench}}{portfolio object representing benchmark, can be NULL}

\item{\code{ac}}{ArcticDB datastore object}

\item{\code{rf}}{risk-free time-series}

\item{\code{col}}{vector of colors for charts, leave NULL for standard DTC colors}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Reporter-new}{\code{Reporter$new()}}
\item \href{#method-Reporter-clean_port_nm}{\code{Reporter$clean_port_nm()}}
\item \href{#method-Reporter-sector_summ}{\code{Reporter$sector_summ()}}
\item \href{#method-Reporter-country_summ}{\code{Reporter$country_summ()}}
\item \href{#method-Reporter-cat_summ}{\code{Reporter$cat_summ()}}
\item \href{#method-Reporter-fina_summ}{\code{Reporter$fina_summ()}}
\item \href{#method-Reporter-ret_combo}{\code{Reporter$ret_combo()}}
\item \href{#method-Reporter-clone}{\code{Reporter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-new}{}}}
\subsection{Method \code{new()}}{
Create Reporter object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$new(port, bench = NULL, rf = NULL, col = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{list of Portfolio Object(s)}

\item{\code{bench}}{benchmark (Portfolio Object), can leave NULL}

\item{\code{rf}}{xts representing risk-free rate, can leave NULL, will default
to BIL ETF}

\item{\code{col}}{vector of colors, leave NULL for standard DTC colors}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-clean_port_nm"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-clean_port_nm}{}}}
\subsection{Method \code{clean_port_nm()}}{
Forces all portfolio names to be unique
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$clean_port_nm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-sector_summ"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-sector_summ}{}}}
\subsection{Method \code{sector_summ()}}{
Sector summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$sector_summ(
  tgt = c("FactsetSector", "GicsMacro", "GicsMap"),
  layer = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tgt}}{target column: FactsetSector, GicsMacro, or GicsMap}

\item{\code{layer}}{1 for security level, 2 for fund level}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
GicsMacro is from piper sandler macro workbook, GicsMap is a mapping
of Factset Sectors to GICS sectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-country_summ"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-country_summ}{}}}
\subsection{Method \code{country_summ()}}{
Country summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$country_summ(tgt = "RiskCountry", layer = 1, rgn = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tgt}}{target column, RiskCountry currently supported}

\item{\code{layer}}{1 for security level, 2 for fund level}

\item{\code{rgn}}{option to include a column with regions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-cat_summ"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-cat_summ}{}}}
\subsection{Method \code{cat_summ()}}{
Summarize category, use sector and country functions to 
  execute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$cat_summ(tbl_cat, tgt_nm, layer = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl_cat}}{data.frame with catagories to summarize}

\item{\code{tgt_nm}}{target column name / category in tbl_cat}

\item{\code{layer}}{1 for security level, 2 for fund level}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-fina_summ"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-fina_summ}{}}}
\subsection{Method \code{fina_summ()}}{
Financial data summary: PE, PB, PFCF, and DY
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$fina_summ(layer = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layer}}{1 for security, 2 for fund}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-ret_combo"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-ret_combo}{}}}
\subsection{Method \code{ret_combo()}}{
Clean and lineup portfolio, benchmark, asset, and rf returns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$ret_combo(freq = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{freq}}{option to change frequency}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
