% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporter.R
\name{Reporter}
\alias{Reporter}
\title{Reporter Object}
\description{
Structure to wrangle portfolios and an optional benchmark into reports
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Reporter-new}{\code{Reporter$new()}}
\item \href{#method-Reporter-clean_port_nm}{\code{Reporter$clean_port_nm()}}
\item \href{#method-Reporter-cat_summ}{\code{Reporter$cat_summ()}}
\item \href{#method-Reporter-sect_summ}{\code{Reporter$sect_summ()}}
\item \href{#method-Reporter-clone}{\code{Reporter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$new(port, bench = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-clean_port_nm"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-clean_port_nm}{}}}
\subsection{Method \code{clean_port_nm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$clean_port_nm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-cat_summ"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-cat_summ}{}}}
\subsection{Method \code{cat_summ()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$cat_summ(tbl_cat, tgt_nm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-sect_summ"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-sect_summ}{}}}
\subsection{Method \code{sect_summ()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$sect_summ(sect_col = c("GicsMacro", "GicsMap", "FactsetSector"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
