% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{unzip_bd_batch}
\alias{unzip_bd_batch}
\title{Unzip Response from BlackDiamond Batch Download}
\usage{
unzip_bd_batch(resp)
}
\arguments{
\item{resp}{response returned by `download_bd_batch`}
}
\value{
json data (list)
}
\description{
Unzip Response from BlackDiamond Batch Download
}
